Data model is now a Store, with initial state moved out and commented
